requirements.txt
Flask==0.12.2
gunicorn==19.7.1
requests==2.9.1 
beautifulsoup4==4.4.1 
nltk==3.2
Flask-pure
python-twitter
pylyrics
billboard.py
flask-wtf